Moved Session to eval package
moved all functions related to evaluation to eval package
