+++ eval/node.go
@@ -1,4 +1,4 @@
-package main
+package eval
 
 import (
 	"reflect"
